+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Replace BootInfo by MemoryMap and offset (#834)
+- Fix sleep::wait timing (#882)
 - Refactor PIC module (#880)
 - Remove dependency on x86_64::instructions::random (#879)
 - Remove dependency on x86_64::instructions::interrupts (#878)